backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
sqlalchemy==2.0.36
python-multipart==0.0.12
pillow==11.0.0
google-generativeai==0.8.3
python-dotenv==1.0.1
pydantic==2.10.2
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.2.1
email-validator==2.1.0
